+++ Sources/SharedModels/TransferUnits/New/Transfer.New.swift
@@ -3,14 +3,14 @@ import Foundation
 public extension Transfer {
   struct New: Hashable, Codable {
     public let info: TransferUnit.Info
-    public let amountValue: Double?
+    public let amountValue: Int?
     public let currencyId: UUID
     public let creditorId: UUID?
     public let borrowerId: UUID?
 
     public init (
       info: TransferUnit.Info,
-      amountValue: Double?,
+      amountValue: Int?,
       currencyId: UUID,
       creditorId: UUID?,
       borrowerId: UUID?
